feat: migrate incident docs functions
